requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.1.4
djangorestframework==3.14.0
exceptiongroup==1.0.4
iniconfig==1.1.1
packaging==22.0
Pillow==9.3.0
pip==21.3.1
pluggy==1.0.0
psycopg2-binary==2.9.5
pytest-django==4.5.2
pytest==7.2.0
python-dotenv==0.21.0
pytz==2022.7
setuptools==60.2.0
sqlparse==0.4.3
tomli==2.0.1
wheel==0.37.1
drf-yasg==1.21.4
